# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WeeklyRun API",
    "description": "WeeklyRun - AI Meeting Orchestration API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/meetings": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "List Meetings",
        "description": "List meetings with optional filters.\n\nP2 Feature: Added meeting_type filter for WEEKLY_REPORT or GENERAL meetings.",
        "operationId": "list_meetings_api_v1_meetings_get",
        "parameters": [
          {
            "name": "team_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Team Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MeetingStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "meeting_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MeetingType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Meeting Type"
            }
          },
          {
            "name": "from_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Date"
            }
          },
          {
            "name": "to_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MeetingResponse"
                  },
                  "title": "Response List Meetings Api V1 Meetings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Create Meeting",
        "description": "Create a new meeting.\n\nP2 Feature: Supports meeting_type for WEEKLY_REPORT or GENERAL meetings.\nGENERAL meetings can have optional agenda_items.",
        "operationId": "create_meeting_api_v1_meetings_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meetings/{meeting_id}": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "Get Meeting",
        "description": "Get a meeting by ID with all related data.",
        "operationId": "get_meeting_api_v1_meetings__meeting_id__get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingWithDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "meetings"
        ],
        "summary": "Delete Meeting",
        "description": "Delete a meeting and all related data.",
        "operationId": "delete_meeting_api_v1_meetings__meeting_id__delete",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meetings/{meeting_id}/status": {
      "patch": {
        "tags": [
          "meetings"
        ],
        "summary": "Update Meeting Status",
        "description": "Update meeting status.",
        "operationId": "update_meeting_status_api_v1_meetings__meeting_id__status_patch",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meetings/{meeting_id}/progress": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "Get Meeting Progress",
        "description": "Get meeting processing progress for polling.\n\nP2 Feature: For GENERAL meetings, weekly_report_loaded step is skipped.",
        "operationId": "get_meeting_progress_api_v1_meetings__meeting_id__progress_get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Meeting Progress Api V1 Meetings  Meeting Id  Progress Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/realtime/meetings/{meeting_id}/progress": {
      "get": {
        "tags": [
          "realtime-meeting"
        ],
        "summary": "Get Meeting Progress",
        "description": "Get current meeting progress (orchestration state).",
        "operationId": "get_meeting_progress_api_v1_realtime_meetings__meeting_id__progress_get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "realtime-meeting"
        ],
        "summary": "Update Meeting Progress",
        "description": "Update meeting progress (current speaker/item).",
        "operationId": "update_meeting_progress_api_v1_realtime_meetings__meeting_id__progress_put",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingProgressUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgendaItem": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "presenter": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Presenter"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 480.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "AgendaItem",
        "description": "Schema for a single agenda item in general meetings."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MeetingCreate": {
        "properties": {
          "team_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team Id"
          },
          "meeting_date": {
            "type": "string",
            "format": "date",
            "title": "Meeting Date"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "meeting_mode": {
            "$ref": "#/components/schemas/MeetingMode",
            "default": "upload"
          },
          "meeting_type": {
            "$ref": "#/components/schemas/MeetingType",
            "default": "weekly_report"
          },
          "agenda_items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AgendaItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agenda Items"
          }
        },
        "type": "object",
        "required": [
          "team_id",
          "meeting_date",
          "title"
        ],
        "title": "MeetingCreate",
        "description": "Schema for creating a meeting."
      },
...
      "MeetingMinutesInfo": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "ai_model": {
            "type": "string",
            "title": "Ai Model"
          },
          "is_edited": {
            "type": "boolean",
            "title": "Is Edited"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ai_model",
          "is_edited",
          "created_at"
        ],
        "title": "MeetingMinutesInfo",
        "description": "Schema for meeting minutes info in meeting response."
      },
      "MeetingMode": {
        "type": "string",
        "enum": [
          "upload",
          "realtime"
        ],
        "title": "MeetingMode",
        "description": "Meeting mode enum."
      },
      "MeetingProgressResponse": {
        "properties": {
          "meeting_id": {
            "type": "string",
            "format": "uuid",
            "title": "Meeting Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "current_speaker_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Speaker Index"
          },
          "current_item_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Item Index"
          },
          "question_tree": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Tree"
          }
        },
        "type": "object",
        "required": [
          "meeting_id",
          "status",
          "current_speaker_index",
          "current_item_index",
          "question_tree"
        ],
        "title": "MeetingProgressResponse",
        "description": "Response with current meeting progress."
      },
      "MeetingProgressUpdate": {
        "properties": {
          "current_speaker_index": {
            "type": "integer",
            "title": "Current Speaker Index"
          },
          "current_item_index": {
            "type": "integer",
            "title": "Current Item Index"
          }
        },
        "type": "object",
        "required": [
          "current_speaker_index",
          "current_item_index"
        ],
        "title": "MeetingProgressUpdate",
        "description": "Request to update meeting progress."
      },
      "MeetingResponse": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "team_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team Id"
          },
          "meeting_date": {
            "type": "string",
            "format": "date",
            "title": "Meeting Date"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "$ref": "#/components/schemas/MeetingStatus"
          },
          "meeting_mode": {
            "type": "string",
            "title": "Meeting Mode"
          },
          "meeting_type": {
            "type": "string",
            "title": "Meeting Type"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "confluence_page_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confluence Page Id"
          },
          "confluence_page_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confluence Page Url"
          },
          "agenda_items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AgendaItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agenda Items"
          }
        },
        "type": "object",
        "required": [
          "created_at",
          "updated_at",
          "id",
          "team_id",
          "meeting_date",
          "title",
          "status",
          "meeting_mode",
          "meeting_type",
          "error_message",
          "confluence_page_id",
          "confluence_page_url",
          "agenda_items"
        ],
        "title": "MeetingResponse",
        "description": "Schema for meeting response."
      },
...
      "MeetingStatus": {
        "type": "string",
        "enum": [
          "created",
          "weekly_report_loaded",
          "preparing",
          "in_progress",
          "recording_done",
          "recording_uploaded",
          "transcribing",
          "transcribed",
          "generating_minutes",
          "draft_ready",
          "published",
          "failed"
        ],
        "title": "MeetingStatus",
        "description": "Meeting status enum."
      },
      "MeetingStatusUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/MeetingStatus"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "MeetingStatusUpdate",
        "description": "Schema for updating meeting status."
      },
      "MeetingType": {
        "type": "string",
        "enum": [
          "weekly_report",
          "general"
        ],
        "title": "MeetingType",
        "description": "Meeting type enum - P2 feature."
      },
      "MeetingWithDetails": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "team_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team Id"
          },
          "meeting_date": {
            "type": "string",
            "format": "date",
            "title": "Meeting Date"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "$ref": "#/components/schemas/MeetingStatus"
          },
          "meeting_mode": {
            "type": "string",
            "title": "Meeting Mode"
          },
          "meeting_type": {
            "type": "string",
            "title": "Meeting Type"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "confluence_page_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confluence Page Id"
          },
          "confluence_page_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confluence Page Url"
          },
          "agenda_items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AgendaItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agenda Items"
          },
          "recording": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecordingInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "weekly_report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WeeklyReportInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "minutes": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MeetingMinutesInfo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "created_at",
          "updated_at",
          "id",
          "team_id",
          "meeting_date",
          "title",
          "status",
          "meeting_mode",
          "meeting_type",
          "error_message",
          "confluence_page_id",
          "confluence_page_url",
          "agenda_items",
          "recording",
          "weekly_report",
          "minutes"
        ],
        "title": "MeetingWithDetails",
        "description": "Schema for meeting with all related data."
      },
...
      "RecordingInfo": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "id",
          "original_filename",
          "file_size",
          "duration_seconds"
        ],
        "title": "RecordingInfo",
        "description": "Schema for recording info in meeting response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WeeklyReportInfo": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "confluence_page_id": {
            "type": "string",
            "title": "Confluence Page Id"
          },
          "confluence_page_url": {
            "type": "string",
            "title": "Confluence Page Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "confluence_page_id",
          "confluence_page_url"
        ],
        "title": "WeeklyReportInfo",
        "description": "Schema for weekly report info in meeting response."
      },
...
    }
  }
}